Reorganize gui classes
