+++ apps/python/ACLIB/source/ui/animation.py
@@ -1,4 +1,4 @@
-from util.util import console
+from util.log import console
 
 
 class Animation:
