+++ apps/python/ACLIB/source/ui/gui/defaults.py
@@ -1,13 +1,13 @@
 import os
 
-from settings import STYLE_DIR, CONFIG_DIR
+from settings import STYLE_DIR, CONFIG_DIR, path
 from storage.config import Config
 
 
 class WidgetConfig:
     @staticmethod
     def load_app_config(app: object, app_name: str):
-        app_config_file = os.path.join(CONFIG_DIR, app_name + '.ini')
+        app_config_file = path(CONFIG_DIR, app_name + '.ini')
         if os.path.isfile(app_config_file):
             config = Config(app_config_file)
 
@@ -34,13 +34,13 @@ class WidgetStyle:
 
             # Try to load from app style config
             if app_name:
-                app_style_file = os.path.join(STYLE_DIR, app_name + '.ini')
+                app_style_file = path(STYLE_DIR, app_name + '.ini')
                 if os.path.isfile(app_style_file):
                     config = Config(app_style_file)
 
             # Try to load from widget style config
             if not config:
-                class_style_file = os.path.join(STYLE_DIR, class_name + '.ini')
+                class_style_file = path(STYLE_DIR, class_name + '.ini')
                 if os.path.isfile(class_style_file):
                     config = Config(class_style_file)
